chore: Improve file layout for SPI classes
